fix: move helper function to tf module
