+++ internal/provider/resource_user.go
@@ -66,20 +66,6 @@ func resourceUser() *schema.Resource {
 	}
 }
 
-func ExpandStringList(configured []interface{}) []string {
-	vs := make([]string, 0, len(configured))
-	for _, v := range configured {
-		val, ok := v.(string)
-		if ok && val != "" {
-			vs = append(vs, val)
-		}
-	}
-	return vs
-}
-func ExpandStringSet(configured *schema.Set) []string {
-	return ExpandStringList(configured.List())
-}
-
 func resourceUserImport(ctx context.Context, d *schema.ResourceData, meta any) ([]*schema.ResourceData, error) {
 	client := meta.(*api.Client)
 	email := d.Id()
@@ -97,7 +83,7 @@ func resourceUserCreate(ctx context.Context, d *schema.ResourceData, meta any) d
 	client := meta.(*api.Client)
 
 	role := d.Get("role").(string)
-	abilities := ExpandStringSet(d.Get("abilities").(*schema.Set))
+	abilities := tf.ExpandStringSet(d.Get("abilities").(*schema.Set))
 
 	if role == "stakeholder" && len(abilities) != 0 {
 		return diag.Errorf("stakeholders cannot have special abilities")
@@ -160,7 +146,7 @@ func resourceUserUpdate(ctx context.Context, d *schema.ResourceData, meta any) d
 	}
 
 	role := d.Get("role").(string)
-	abilities := ExpandStringSet(d.Get("abilities").(*schema.Set))
+	abilities := tf.ExpandStringSet(d.Get("abilities").(*schema.Set))
 
 	if role == "stakeholder" && len(abilities) != 0 {
 		return diag.Errorf("stakeholders cannot have special abilities")
